Include native version of reverse for InlineStrings
This is the first in a series of PRs where I'm going to attempt to
provide "native" string functions that operate on and return
InlineStrings. The advantage is that when working with InlineStrings,
the normal set of string functions won't "poison" your data by
unnecessarily introducing `String`s, but will operate directly on the
InlineString type and return the same type when possible.

The `reverse` function is a good example of that: it can always return
the same type as the input, and in my benchmarks, it can be 2-4x faster
than the "default" `String` method.