Ensure shuffled bits get cleared out in "modifying" operations (#33)
* Ensure shuffled bits get cleared out in "modifying" operations

Fixes #32. The core issue here is we're taking a few shortcuts in some operations
like chop, chomp, first, last where we just shuffle the bits around and OR the
new length. The problem is there can be "extra bits" in the inline string that
can then affect operations like `==`. So we need to ensure in these optimized
"modifying" operations, these extra bits get zeroed out to ensure a consistent
bit representation.

* remove old code

* fix tests